Factor mailbox controller
Separate out an identifiable controller, with constructor
taking any suitable pre-informers and clients.

Also fix some oversights in the doc.
